% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esalqmet_stations.R
\docType{data}
\name{esalqmet_conventional}
\alias{esalqmet_conventional}
\title{Conventional station weather data}
\value{
\item{esalqmet_conventional}{a tibble}
}
\description{
Conventional station weather data
}
